+++ examples/all-the-things.rs
@@ -1,5 +1,7 @@
 #![allow(clippy::from_over_into, clippy::disallowed_names)]
 
+use std::error::Error as _;
+
 #[derive(Debug, Clone)]
 struct Foo(String);
 
@@ -36,13 +38,22 @@ struct Baz<T> {
 
 impl_more::forward_display!(<T> in Baz<T> => inner);
 
+#[derive(Debug)]
+enum Err {
+    Io(std::io::Error),
+    Generic(String),
+}
+
+impl_more::impl_display_enum!(Err, Io(err) => "{err}", Generic(msg) => "{msg}");
+impl_more::impl_error_enum!(Err, Io(err) => err);
+
 fn main() {
     let mut foo = Foo("bar".to_owned());
 
     foo.as_mut().push('!');
     assert_eq!(foo.as_ref(), "bar!");
 
-    assert_eq!(foo.len(), 3);
+    assert_eq!(foo.len(), 4);
 
     let mut foo = Foo("bar".to_owned());
     foo.push('!');
@@ -55,4 +66,8 @@ fn main() {
 
     assert_eq!(FooEnum::Bar.to_string(), "bar");
     assert_eq!(FooEnum::Qux.to_string(), "qux");
+
+    let io_err = std::io::Error::new(std::io::ErrorKind::Other, "test");
+    assert!(Err::Io(io_err).source().is_some());
+    assert!(Err::Generic("oops".to_owned()).source().is_none());
 }