+++ src/error.rs
@@ -1,20 +1,25 @@
 /// Implements [`Error`] for enums.
 ///
+/// Emitted code is not compatible with `#[no_std]`.
+///
 /// # Examples
 ///
 /// ```
+/// # extern crate alloc;
+/// use std::error::Error as _;
+///
 /// #[derive(Debug)]
-/// enum Foo {
-///     Bar,
-///     Qux(std::io::Error),
+/// enum Err {
+///     Io(std::io::Error),
+///     Generic(String),
 /// }
 ///
-/// impl_more::impl_display_enum!(Foo, Bar => "bar", Qux => "qux");
-/// impl_more::impl_error_enum!(Foo, Qux(err) => err);
+/// impl_more::impl_display_enum!(Err, Io(err) => "{err}", Generic(msg) => "{msg}");
+/// impl_more::impl_error_enum!(Err, Io(err) => err);
 ///
 /// # let io_err = std::io::Error::new(std::io::ErrorKind::Other, "test");
-/// assert_eq!(Foo::Bar.source().is_none());
-/// assert_eq!(Foo::Qux(io_err).source().is_some());
+/// assert!(Err::Io(io_err).source().is_some());
+/// assert!(Err::Generic("oops".to_owned()).source().is_none());
 /// ```
 ///
 /// [`Error`]: std::error::Error
@@ -34,7 +39,7 @@ macro_rules! impl_error_enum {
     };
 
     ($ty:ty, $($variant:ident ($($inner:ident),+) => $source:expr),+) => {
-        impl_error_enum!($ty, $($variant ($($inner),+) => $source),+ ,)
+        $crate::impl_error_enum!($ty, $($variant ($($inner),+) => $source),+ ,);
     };
 
     ($ty:ty, $($variant:ident { $($inner:ident),+ } => $source:expr),+ ,) => {
@@ -51,7 +56,7 @@ macro_rules! impl_error_enum {
     };
 
     ($ty:ty, $($variant:ident { $($inner:ident),+ } => $source:expr),+) => {
-        impl_error_enum!($ty, $($variant { $($inner),+ } => $source),+ ,)
+        $crate::impl_error_enum!($ty, $($variant { $($inner),+ } => $source),+ ,);
     };
 
     ($ty:ty,) => {
@@ -59,12 +64,13 @@ macro_rules! impl_error_enum {
     };
 
     ($ty:ty) => {
-        impl_error_enum!($ty,)
+        $crate::impl_error_enum!($ty,);
     };
 }
 
 #[cfg(test)]
 mod tests {
+    use alloc::string::String;
     use std::error::Error as _;
 
     #[test]