+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Unreleased
 
-- Add basic `Error` derive for enums behind new on-by-default crate feature `std`.
-- Fix `#[no_std]` support.
+- Add basic `Error` derive for enums.
 
 ## 0.1.3 - 2023-02-24